refactor related collections use relatedCollections as a top level prop on a collection
